Adding concerts into the view model and initial view for concert
