--- visionOS-tutorial/visionOS-tutorial/Data/Playlist.swift
@@ -1,16 +1,0 @@
-//
-//  Playlist.swift
-//  VisionProFirstProject
-//
-//  Created by Hollins, Cecilia on 28/6/2023.
-//
-
-import SwiftUI
-
-struct Playlist: Identifiable, Hashable {
-    let id = UUID()
-    let title: String
-    let description: String
-    let coverImage: String
-    let songs: [Song]
-}
