--- visionOS-tutorial/visionOS-tutorial/Data/Song.swift
@@ -1,15 +1,0 @@
-//
-//  Song.swift
-//  VisionProFirstProject
-//
-//  Created by Hollins, Cecilia on 28/6/2023.
-//
-
-import SwiftUI
-
-struct Song: Hashable {
-    let songTitle: String
-    let artist: String
-    let albumCover: String
-    let songLength: String
-}
