+++ visionOS-tutorial/visionOS-tutorial/Model/Song.swift
@@ -1,0 +1,15 @@
+//
+//  Song.swift
+//  VisionProFirstProject
+//
+//  Created by Hollins, Cecilia on 28/6/2023.
+//
+
+import SwiftUI
+
+struct Song: Hashable {
+    let songTitle: String
+    let artist: String
+    let albumCover: String
+    let songLength: String
+}
